+++ divio_cli/cloud.py
@@ -140,6 +140,44 @@ class CloudClient(object):
         request = api_requests.ProjectListRequest(self.session)
         return request()
 
+    def get_services(self, region_uuid=None, application_uuid=None):
+        kwargs = {}
+
+        # TODO this smells like a security issue
+        kwargs["filter_region"] = (
+            f"region={region_uuid}" if region_uuid else ""
+        )
+
+        kwargs["filter_website"] = (
+            f"website={application_uuid}" if application_uuid else ""
+        )
+        request = api_requests.ListServicesRequest(
+            self.session,
+            url_kwargs=kwargs,
+        )
+        return request()
+
+    def get_service_instances(self, environment_uuid):
+        request = api_requests.ListServiceInstancesRequest(
+            self.session,
+            url_kwargs={"environment_uuid": environment_uuid},
+        )
+        return request()
+
+    def add_service_instances(
+        self, environment_uuid, prefix, region_uuid, service_uuid
+    ):
+        request = api_requests.CreateServiceInstanceRequest(
+            self.session,
+            data={
+                "environment": environment_uuid,
+                "region": region_uuid,
+                "service": service_uuid,
+                "prefix": prefix,
+            },
+        )
+        return request()
+
     def ssh(self, website_id, environment):
         project_data = self.get_project(website_id)
         try:
@@ -740,7 +778,7 @@ class CloudClient(object):
             results, _ = json_response_request_paginate(
                 api_requests.ListServiceInstancesRequest,
                 self.session,
-                params={"environment": environment_uuid},
+                url_kwargs={"environment_uuid": environment_uuid},
                 limit_results=limit_results,
             )
 
@@ -914,6 +952,14 @@ class CloudClient(object):
                 "backup_restore_uuid": backup_restore_uuid,
             },
         )()
+
+    def get_regions(self):
+        request = api_requests.ListRegionsRequest(self.session)
+        return request()
+
+    def get_organisations(self):
+        request = api_requests.ListOrganisationsRequest(self.session)
+        return request()
 
 
 class WritableNetRC(netrc):