+++ divio_cli/test/test_commands.py
@@ -29,9 +29,12 @@ TEST_COMMANDS_CLICK = [
     ["app", "logs", "test"],
     ["app", "status"],
     ["app", "update"],
+    ["app", "service-instances", "list"],
     ["version"],
     ["version", "-s"],
     ["version", "-m"],
+    ["regions", "list"],
+    ["organisations", "list"],
 ]