+++ CHANGELOG.rst
@@ -5,6 +5,9 @@ x.x.x (x-x-x)
 -------------------
 
 * Added tests for environment variables and deployments commands.
+* Added support to list and add services to an application.
+* Added support to list regions.
+* Added support to list organisations.
 
 3.17.2 (2023-03-16)
 -------------------